+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## 1.4.1 - 2016-10-04
 ### Changed
 - Register session_write_close as a shutdown function when using sessions.